new version of the deploy control, not tested yet

+++ mconf-bbb/recipes/deploy.rb
@@ -19,11 +19,12 @@ execute "unzip_bigbluebutton" do
   user "root"
   command "unzip -o -d #{node[:mconf][:bbb][:deploy_dir]}/#{node[:mconf][:bbb][:version]} -q #{Chef::Config[:file_cache_path]}/#{node[:mconf][:bbb][:file]}"
   action :run
+  only_if do File.exists?('#{node[:mconf][:bbb][:deploy_dir]}/deploy_needed') end
 end
 
 timestamp = Time.new.strftime("%Y%m%d-%H%M%S")
-#backup_dir = "#{node[:mconf][:bbb][:deploy_dir]}/backup-#{timestamp}"
-backup_dir = "/var/mconf/deploy/mconf-bbb/backup-20120716-155448"
+backup_dir = "#{node[:mconf][:bbb][:deploy_dir]}/backup-#{timestamp}"
+#backup_dir = "/var/mconf/deploy/mconf-bbb/backup-20120716-155448"
 
 node[:bbb][:modules].each do |name|
   module_deploy_dir = node["bbb"]["#{name}"]["deploy_dir"]
@@ -32,6 +33,7 @@ node[:bbb][:modules].each do |name|
   directory "#{backup_dir}/#{name}" do
     recursive true
     action :create
+    only_if do File.exists?('#{node[:mconf][:bbb][:deploy_dir]}/deploy_needed') end
   end
   
   # backup only the bbb related apps on /usr/local/bin/
@@ -45,6 +47,7 @@ node[:bbb][:modules].each do |name|
     user "root"
     command "cp -r #{module_deploy_dir}/#{backup_files} #{module_backup_dir}"
     action :run
+    only_if do File.exists?('#{node[:mconf][:bbb][:deploy_dir]}/deploy_needed') end
   end
   
   execute "deploy_module" do
@@ -52,6 +55,18 @@ node[:bbb][:modules].each do |name|
     cwd "#{node[:mconf][:bbb][:deploy_dir]}"
     command "cp -r #{node[:mconf][:bbb][:version]}/#{name}/* #{module_deploy_dir}"
     action :run
+    only_if do File.exists?('#{node[:mconf][:bbb][:deploy_dir]}/deploy_needed') end
   end
+end
+
+#register deployed version
+file "#{node[:mconf][:bbb][:deploy_dir]}/.deployed" do
+  action :create
+  content "#{node[:mconf][:bbb][:version]}"
+end
+
+#delete deploy flag after deployement
+file "#{node[:mconf][:bbb][:deploy_dir]}/deploy_needed" do
+  action :delete
 end
